requirements.txt
pandas
watchdog
streamlit
plotly
pyyaml
openpyxl
xlrd
mcp
